requirements.txt
logging==0.4.9.6
mysql-connector-python==9.5.0
